Some test suite updates.
Add tst-key4 from libc, although we need to deal with 1 less pthread key
than normal due to something that's going on with pthread when libeatmydata
is present that I don't quite know what's going on.

We also update tst-cancel and the test run script so that we better capture
errors from the test suite.